feat: real implementation pass + rebrand to discord_rich_presence
- Discord Application registered: 1507704875939790889 (replaces 18-zero placeholder)
- Rename DLL/INIs to lowercase discord_rich_presence.{dll,ini,_translation.ini}
- PlayerSnapshot: real HP via GetPermanentActorValue/GetActorValue against ActorValue::health
- CombatTracker: real polling of IsInCombat() + currentCombatTarget.get() for target name
- EventTracker: live BSTEventSink for TerminalHacked + Workshop::ItemPlacedEvent (1500ms debounce)
- vcpkg pinned fmt to 10.2.1 (fmt 11+ breaks REL::Version formatter)
- VCPKG_VISUAL_STUDIO_PATH locked to VS 2022 BuildTools (MSVC 14.44) to avoid CommonLibF4 hkVector4 break on 14.50
- E2E integration test: full handshake + SetActivity against real AppID, validates Discord ACK
- 39 unit/contract + 4 integration tests all green (18 assertions live Discord)
- DLL: 790528 bytes, exports F4SEPlugin_Load + F4SEPlugin_Version, deployed

+++ src/Config/Settings.cpp
@@ -91,7 +91,8 @@ Settings parseFromString(const std::string& iniContents)
     s.showName = parseBool(ini.GetValue(main.c_str(), Constants::IniKey::kShowName.data()), s.showName);
     s.showLvl = parseBool(ini.GetValue(main.c_str(), Constants::IniKey::kShowLvl.data()), s.showLvl);
     s.showCaps = parseBool(ini.GetValue(main.c_str(), Constants::IniKey::kShowCaps.data()), s.showCaps);
-    s.maxCapsToShow = ini.GetLongValue(main.c_str(), Constants::IniKey::kMaxCapsToShow.data(), s.maxCapsToShow);
+    s.maxCapsToShow =
+        ini.GetLongValue(main.c_str(), Constants::IniKey::kMaxCapsToShow.data(), static_cast<long>(s.maxCapsToShow));
     s.showHp = parseBool(ini.GetValue(main.c_str(), Constants::IniKey::kShowHp.data()), s.showHp);
     s.showEventStatuses =
         parseBool(ini.GetValue(main.c_str(), Constants::IniKey::kShowEventStatuses.data()), s.showEventStatuses);
